requirements.txt
Flask==3.0.0
flask-mysqldb==0.2.0
mysqlclient==2.1.1
requests==2.31.0
env==0.1.0
python-dotenv==1.0.1